32-bit support (#19)

+++ examples/alignment.cpp
@@ -16,7 +16,6 @@
 #include "tmc/ex_cpu.hpp"
 #include "tmc/spawn_many.hpp"
 
-#include <cinttypes>
 #include <cstdio>
 #include <ranges>
 
@@ -32,18 +31,17 @@ struct alignas(ALIGNMENT) aligned_struct {
 };
 
 static void check_alignment(void* ptr) {
-  auto low_bits = reinterpret_cast<uint64_t>(ptr) % ALIGNMENT;
+  auto low_bits = reinterpret_cast<size_t>(ptr) % ALIGNMENT;
   if (low_bits != 0) {
     std::printf(
-      "FAIL: Expected align %" PRIu64 " but got align %" PRIu64 "\n", ALIGNMENT,
-      low_bits
+      "FAIL: Expected align %zu but got align %zu\n", ALIGNMENT, low_bits
     );
     std::fflush(stdout);
   }
 }
 static task<void> run_one(int i, unaligned_struct* ur, aligned_struct* ar) {
   static_assert(alignof(aligned_struct) == ALIGNMENT);
-  static_assert(sizeof(void*) == sizeof(uint64_t));
+  static_assert(sizeof(void*) == sizeof(size_t));
   unaligned_struct u;
   aligned_struct a;
   u.value = i & 0xFF;
