+++ examples/callback_awaitable.hpp
@@ -22,8 +22,9 @@
 
 #pragma once
 
-#include "tmc/detail/concepts.hpp"
-#include "tmc/detail/thread_locals.hpp"
+#include "tmc/detail/compat.hpp"   // for TMC_FORCE_INLINE
+#include "tmc/detail/concepts.hpp" // for result_storage_t, awaitable_traits
+#include "tmc/detail/thread_locals.hpp" // for this_thread
 #include "tmc/task.hpp"
 
 #include <coroutine>
@@ -202,7 +203,7 @@ template <IsAwCallback Awaitable> struct awaitable_traits<Awaitable> {
     awaitable.customizer.done_count = DoneCount;
   }
 
-  static void set_flags(self_type& awaitable, uint64_t Flags) {
+  static void set_flags(self_type& awaitable, size_t Flags) {
     awaitable.customizer.flags = Flags;
   }
 };
